Crate xorshift

Implementation of the high performance xoroshiro128+, xorshift128+, xorshift1024*, and splitmix64 pseudo random number generators.

Implements the Rand, Rng, and SeedableRng traits from the rand crate.

Usage

[dependencies]
xorshift = "0.1"

extern crate xorshift;

Examples

Reexports

pub use splitmix64::SplitMix64;

pub use xoroshiro128::Xoroshiro128;

pub use xorshift128::Xorshift128;

pub use xorshift1024::Xorshift1024;

Modules

splitmix64
The SplitMix64 random number generator.

xoroshiro128
The Xoroshiro128+ random number generator.

xorshift1024
The Xorshift1024* random number generator.

xorshift128
The Xorshift128+ random number generator.

Traits

Rand
A type that can be randomly generated using an Rng.

Rng
A random number generator.

RngJump
A random number generator with jumpable state.

SeedableRng
A random number generator that can be explicitly seeded to produce the same stream of randomness multiple times.